+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/primitive/element/ProtobufEnumDefinition.kt
@@ -1,11 +1,12 @@
 package io.kanro.idea.plugin.protobuf.lang.psi.primitive.element
 
 import io.kanro.idea.plugin.protobuf.Icons
+import io.kanro.idea.plugin.protobuf.lang.psi.primitive.stratify.ProtobufOptionOwner
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.structure.ProtobufDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.structure.ProtobufNumberScope
 import javax.swing.Icon
 
-interface ProtobufEnumDefinition : ProtobufNumberScope, ProtobufDefinition {
+interface ProtobufEnumDefinition : ProtobufNumberScope, ProtobufDefinition, ProtobufOptionOwner {
     @JvmDefault
     override fun type(): String {
         return "enum"
@@ -14,5 +15,10 @@ interface ProtobufEnumDefinition : ProtobufNumberScope, ProtobufDefinition {
     @JvmDefault
     override fun getIcon(unused: Boolean): Icon? {
         return Icons.ENUM
+    }
+
+    @JvmDefault
+    override fun allowAlias(): Boolean {
+        return this.options("allow_alias").firstOrNull()?.value()?.booleanValue?.text == "true"
     }
 }