+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/primitive/structure/ProtobufNumberScope.kt
@@ -22,4 +22,9 @@ interface ProtobufNumberScope : ProtobufScope {
             findChildren()
         }
     }
+
+    @JvmDefault
+    fun allowAlias(): Boolean {
+        return false
+    }
 }